Optimize Player Full Karma querying and update DTO
The method for querying a player's full karma has been optimized by compiling the EF query. Compiling the query improves the performance as it removes the overhead of query compilation at runtime. This change ensures that the application is more performant and responsive to the end user.

Moreover, the AccountFullKarmaDTO model was updated to better reflect the structure of the Karma system, enhancing readability and maintainability of the code.

Signed-off-by: Sakura Akeno Isayeki <sakura.isayeki@nodsoft.net>