feat(migrations): Add Nullable context migrations
This commit adds nullable context migrations to the codebase. The `UpdateNullableContext` migration updates several columns in different tables to be non-nullable. Specifically, it modifies the `Players`, `ChatMessages`, and `BlobName` columns in the `Replays` table, the `Reason` column in the `PostModActions` table, and the `Username`, `OptOutChanged`, `Tag`, `Name`, and `Description` columns in the `Clans` table. These changes ensure that these columns cannot have null values.