fix(api/posts/sent): Include replay in post queries
This commit fixes an issue where the minimap state  was not attached to posts, in queries for sent posts. The code changes include adding the `Include` method to include the replay in the query for both cases. This ensures that the replay is now properly retrieved along with other relevant data when querying posts by player or author.