chore(web): Comment out global PSA
The global PSA message has been commented out in the code. This change removes the display of a success alert and information alert on the web page. The success alert previously informed users about the deployment of a replay RCE detection system, while the information alert provided details about a Replay Scanner app. These alerts are no longer visible on the platform.

This commit does not introduce any functional changes but modifies the HTML file to comment out these sections of code.