build: [Experiment] Enable ReadyToRun publishing
This commit enables the ReadyToRun publishing feature in the WowsKarma.Api project. By setting the `<PublishReadyToRun>` property to `true`, we can now generate and publish ReadyToRun images, which can improve startup performance of the application. This is an experimental change aimed at optimizing the build process and enhancing overall performance.