feat(Post): Add Customer Support Ticket ID property on Post entity
This commit adds a new property called "CustomerSupportTicketId" to the Post model. This property is used to store the Customer Support ticket ID associated with a post. It can take a positive integer value up to 9 digits or be set as null if no ticket is associated. The purpose of this property is to allow WG staff to track and manage player reports.

Implements database requirements for #165.